Massive update to code
The plugin is now backed by SQL, has a clipboard  for the current edit,
and is saving somewhat.  Its a bit riddled with NPEs at the moment but
they should disappear quickly.

Made it where there are up to 4 lines of news story body, plus one line
for the headline.

Also, added a third command.  /news will show the headlines and article
numbers, /readnews # shows the full article.